feat(phase-1.1): decompose api/server.py into 8+ routers (#896)
Extracted all route handlers from the monolithic server.py (2490 lines)
into 10 focused route modules under maxwell_daemon/api/routes/, reducing
server.py to 325 lines.

New files:
- maxwell_daemon/api/deps.py        — auth dependency factories
- maxwell_daemon/api/routes/actions.py
- maxwell_daemon/api/routes/audit.py
- maxwell_daemon/api/routes/backends.py
- maxwell_daemon/api/routes/dispatch.py
- maxwell_daemon/api/routes/events.py
- maxwell_daemon/api/routes/fleet.py
- maxwell_daemon/api/routes/issues.py
- maxwell_daemon/api/routes/ssh.py
- maxwell_daemon/api/routes/webhooks.py
- maxwell_daemon/api/routes/work_items.py

Updated tests to patch the correct route module attributes
(registry in routes/backends.py, log in routes/webhooks.py)
rather than the now-thin server.py.

All 2553 unit tests pass. Stable operator contract unchanged.
